Port to Solaris 11 / Solaris Studio 12.3 (== CC 5.12) Fix to BasicStream.h (unitialized data member) that caused occasional failures in Ice/stream test

+++ config/makeprops.py
@@ -75,9 +75,9 @@ struct PropertyArray
     const Property* properties;
     const int length;
 
-    PropertyArray(const Property* p, int len) :
+    PropertyArray(const Property* p, size_t len) :
         properties(p),
-        length(len)
+        length(static_cast<int>(len))
     {
     }
 };
